# Implement DIFF, DIFF1, ANDOR and ONE for BITOP