--- a/src/bitops.c
+++ b/src/bitops.c
@@ -9,6 +9,19 @@
 
 #include "server.h"
 
+#ifdef HAVE_AVX2
+/* Define __MM_MALLOC_H to prevent importing the memory aligned
+ * allocation functions, which we don't use. */
+#define __MM_MALLOC_H
+#include <immintrin.h>
+#endif
+
+#ifdef HAVE_AVX2
+#define BITOP_USE_AVX2 (__builtin_cpu_supports("avx2"))
+#else
+#define BITOP_USE_AVX2 0
+#endif
+
 /* -----------------------------------------------------------------------------
  * Helpers and low level bit functions.
  * -------------------------------------------------------------------------- */
@@ -637,6 +650,164 @@ void getbitCommand(client *c) {
     addReply(c, bitval ? shared.cone : shared.czero);
 }
 
+#ifdef HAVE_AVX2
+/* Compute the given bitop operation using AVX2 intrinsics.
+ * Return how many bytes were successfully processed, as AVX2 operates on
+ * 256-bit registers so if `minlen` is not a multiple of 32 some of the bytes
+ * will be skipped. They will be taken care for in the unoptimized loop in the
+ * main bitopCommand function. */
+ATTRIBUTE_TARGET_AVX2
+unsigned long bitopCommandAVX(unsigned char **keys, unsigned char *res, 
+                              unsigned long op, unsigned long numkeys,
+                              unsigned long minlen) {
+    const unsigned long step = sizeof(__m256i);
+
+    unsigned long i;
+    unsigned long processed = 0;
+    unsigned char *res_start = res;
+    unsigned char *fst_key = keys[0];
+    // TODO: this is debug code, remove.
+    char *bitop_disable_avx_env = getenv("BITOP_DISABLE_AVX");
+    if (bitop_disable_avx_env && !strncmp(bitop_disable_avx_env, "1", 1)) {
+        return 0;
+    }
+
+    if (minlen < step) {
+        return 0;
+    }
+
+    if (op != BITOP_DIFF && op != BITOP_DIFF1 && op != BITOP_ANDOR) {
+        memcpy(res, keys[0], minlen);
+    }
+
+    const __m256i max256 = _mm256_set1_epi64x(-1);
+    const __m256i zero256 = _mm256_set1_epi64x(0);
+
+    switch (op) {
+    case BITOP_AND:
+        while (minlen >= step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+
+            for (i = 1; i < numkeys; i++) {
+                __m256i lkey = _mm256_lddqu_si256((__m256i*)(keys[i]+processed));
+                lres = _mm256_and_si256(lres, lkey);
+            }
+            _mm256_storeu_si256((__m256i*)res, lres);
+            res += step;
+            processed += step;
+            minlen -= step;
+        }
+        break;
+    case BITOP_DIFF:
+    case BITOP_DIFF1:
+    case BITOP_ANDOR:
+    case BITOP_OR:
+        while (minlen >= step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+
+            for (i = 1; i < numkeys; i++) {
+                __m256i lkey = _mm256_lddqu_si256((__m256i*)(keys[i]+processed));
+                lres = _mm256_or_si256(lres, lkey);
+            }
+            _mm256_storeu_si256((__m256i*)res, lres);
+            res += step;
+            processed += step;
+            minlen -= step;
+        }
+        break;
+    case BITOP_XOR:
+        while (minlen >= step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+
+            for (i = 1; i < numkeys; i++) {
+                __m256i lkey = _mm256_lddqu_si256((__m256i*)(keys[i]+processed));
+                lres = _mm256_xor_si256(lres, lkey);
+            }
+            _mm256_storeu_si256((__m256i*)res, lres);
+            res += step;
+            processed += step;
+            minlen -= step;
+        }
+        break;
+    case BITOP_NOT:
+        while (minlen >= step) {
+             __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+            lres = _mm256_xor_si256(lres, max256);
+            _mm256_storeu_si256((__m256i*)res, lres);
+            res += step;
+            processed += step;
+            minlen -= step;
+        }
+        break;
+    case BITOP_ONE:
+	while (minlen >= step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+            __m256i common_bits = zero256;
+
+            for (i = 1; i < numkeys; i++) {
+                __m256i lkey = _mm256_lddqu_si256((__m256i*)(keys[i]+processed));
+                __m256i common = _mm256_and_si256(lres, lkey);
+                common_bits = _mm256_or_si256(common_bits, common);
+
+                lres = _mm256_xor_si256(lres, lkey);
+            }
+            lres = _mm256_andnot_si256(common_bits, lres);
+            _mm256_storeu_si256((__m256i*)res, lres);
+            res += step;
+            processed += step;
+            minlen -= step;
+        }
+        break;
+    default:
+        break;
+    }
+
+    res = res_start;
+    switch (op) {
+    case BITOP_DIFF:
+        for (i = 0; i < processed; i += step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+            __m256i fkey = _mm256_lddqu_si256((__m256i*)fst_key);
+
+            lres = _mm256_andnot_si256(lres, fkey);
+            _mm256_storeu_si256((__m256i*)res, lres);
+
+            res += step;
+            fst_key += step;
+        }
+        break;
+    case BITOP_DIFF1:
+        for (i = 0; i < processed; i += step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+            __m256i fkey = _mm256_lddqu_si256((__m256i*)fst_key);
+
+            lres = _mm256_andnot_si256(fkey, lres);
+            _mm256_storeu_si256((__m256i*)res, lres);
+
+            res += step;
+            fst_key += step;
+        }
+        break;
+    case BITOP_ANDOR:
+        for (i = 0; i < processed; i += step) {
+            __m256i lres = _mm256_lddqu_si256((__m256i*)res);
+            __m256i fkey = _mm256_lddqu_si256((__m256i*)fst_key);
+
+            lres = _mm256_and_si256(fkey, lres);
+            _mm256_storeu_si256((__m256i*)res, lres);
+
+            res += step;
+            fst_key += step;
+        }
+        break;
+    default:
+        break;
+    }
+
+    return processed;
+}
+#endif // HAVE_AVX2
+
 /* BITOP op_name target_key src_key1 src_key2 src_key3 ... src_keyN */
 REDIS_NO_SANITIZE("alignment")
 void bitopCommand(client *c) {
@@ -723,23 +894,31 @@ void bitopCommand(client *c) {
         unsigned char output, byte, disjunction, common_bits;
         unsigned long i;
 
-        /* Fast path: as far as we have data for all the input bitmaps we
+        j = 0;
+
+#if defined(HAVE_AVX2)
+        if (BITOP_USE_AVX2) {
+            j = bitopCommandAVX(src, res, op, numkeys, minlen);
+            minlen -= j;
+        }
+#endif
+
+#if !defined(USE_ALIGNED_ACCESS)
+        /* We don't have AVX2 but we still have fast path:
+         * as far as we have data for all the input bitmaps we
          * can take a fast path that performs much better than the
          * vanilla algorithm. On ARM we skip the fast path since it will
          * result in GCC compiling the code using multiple-words load/store
          * operations that are not supported even in ARM >= v6. */
-        j = 0;
-        #ifndef USE_ALIGNED_ACCESS
-        if (minlen >= sizeof(unsigned long)*4 && numkeys <= 16) {
-            unsigned long *lp[16];
+        if (minlen >= sizeof(unsigned long)*4) {
+            unsigned long **lp = (unsigned long**)src;
             unsigned long *lres = (unsigned long*) res;
             unsigned long *first_key = (unsigned long*)src[0];
             unsigned long lcommon_bits[4];
 
-            size_t step = sizeof(unsigned long)*4;
+            const size_t step = sizeof(unsigned long)*4;
             size_t processed = 0;
-
-            memcpy(lp,src,sizeof(unsigned long*)*numkeys);
+            size_t k = 0;
 
             if (op != BITOP_DIFF && op != BITOP_DIFF1 && op != BITOP_ANDOR)
                 memcpy(lres,src[0],minlen);
@@ -748,12 +927,12 @@ void bitopCommand(client *c) {
             if (op == BITOP_AND) {
                 while(minlen >= step) {
                     for (i = 1; i < numkeys; i++) {
-                        lres[0] &= lp[i][0];
-                        lres[1] &= lp[i][1];
-                        lres[2] &= lp[i][2];
-                        lres[3] &= lp[i][3];
-                        lp[i] += 4;
+                        lres[0] &= lp[i][k+0];
+                        lres[1] &= lp[i][k+1];
+                        lres[2] &= lp[i][k+2];
+                        lres[3] &= lp[i][k+3];
                     }
+                    k += 4;
                     lres += 4;
                     j += step;
                     minlen -= step;
@@ -761,12 +940,12 @@ void bitopCommand(client *c) {
             } else if (op == BITOP_OR) {
                 while(minlen >= step) {
                     for (i = 1; i < numkeys; i++) {
-                        lres[0] |= lp[i][0];
-                        lres[1] |= lp[i][1];
-                        lres[2] |= lp[i][2];
-                        lres[3] |= lp[i][3];
-                        lp[i]+=4;
+                        lres[0] |= lp[i][k+0];
+                        lres[1] |= lp[i][k+1];
+                        lres[2] |= lp[i][k+2];
+                        lres[3] |= lp[i][k+3];
                     }
+                    k += 4;
                     lres += 4;
                     j += step;
                     minlen -= step;
@@ -774,12 +953,12 @@ void bitopCommand(client *c) {
             } else if (op == BITOP_XOR) {
                 while(minlen >= step) {
                     for (i = 1; i < numkeys; i++) {
-                        lres[0] ^= lp[i][0];
-                        lres[1] ^= lp[i][1];
-                        lres[2] ^= lp[i][2];
-                        lres[3] ^= lp[i][3];
-                        lp[i]+=4;
+                        lres[0] ^= lp[i][k+0];
+                        lres[1] ^= lp[i][k+1];
+                        lres[2] ^= lp[i][k+2];
+                        lres[3] ^= lp[i][k+3];
                     }
+                    k += 4;
                     lres += 4;
                     j += step;
                     minlen -= step;
@@ -797,12 +976,12 @@ void bitopCommand(client *c) {
             } else if (op == BITOP_DIFF || op == BITOP_DIFF1 || op == BITOP_ANDOR) {
                 while(minlen >= step) {
                     for (i = 1; i < numkeys; i++) {
-                        lres[0] |= lp[i][0];
-                        lres[1] |= lp[i][1];
-                        lres[2] |= lp[i][2];
-                        lres[3] |= lp[i][3];
-                        lp[i] += 4;
+                        lres[0] |= lp[i][k+0];
+                        lres[1] |= lp[i][k+1];
+                        lres[2] |= lp[i][k+2];
+                        lres[3] |= lp[i][k+3];
                     }
+                    k += 4;
                     lres += 4;
                     j += step;
                     minlen -= step;
@@ -847,30 +1026,30 @@ void bitopCommand(client *c) {
                     memset(lcommon_bits, 0, sizeof(unsigned long)*4);
 
                     for (i = 1; i < numkeys; i++) {
-                        lcommon_bits[0] |= (lres[0] & lp[i][0]);
-                        lcommon_bits[1] |= (lres[1] & lp[i][1]);
-                        lcommon_bits[2] |= (lres[2] & lp[i][2]);
-                        lcommon_bits[3] |= (lres[3] & lp[i][3]);
+                        lcommon_bits[0] |= (lres[0] & lp[i][k+0]);
+                        lcommon_bits[1] |= (lres[1] & lp[i][k+1]);
+                        lcommon_bits[2] |= (lres[2] & lp[i][k+2]);
+                        lcommon_bits[3] |= (lres[3] & lp[i][k+3]);
 
-                        lres[0] ^= lp[i][0];
-                        lres[1] ^= lp[i][1];
-                        lres[2] ^= lp[i][2];
-                        lres[3] ^= lp[i][3];
-
-                        lres[0] &= ~lcommon_bits[0];
-                        lres[1] &= ~lcommon_bits[1];
-                        lres[2] &= ~lcommon_bits[2];
-                        lres[3] &= ~lcommon_bits[3];
-
-                        lp[i] += 4;
+                        lres[0] ^= lp[i][k+0];
+                        lres[1] ^= lp[i][k+1];
+                        lres[2] ^= lp[i][k+2];
+                        lres[3] ^= lp[i][k+3];
                     }
+
+                    lres[0] &= ~lcommon_bits[0];
+                    lres[1] &= ~lcommon_bits[1];
+                    lres[2] &= ~lcommon_bits[2];
+                    lres[3] &= ~lcommon_bits[3];
+
+                    k += 4;
                     lres += 4;
                     j += step;
                     minlen -= step;
                 }
             }
         }
-        #endif
+#endif // !defined(USE_ALIGNED_ACCESS)
 
         /* j is set to the next byte to process by the previous loop. */
         for (; j < maxlen; j++) {