… src/commands.def
… src/commands/bitop.json
… tests/unit/bitops.tcl