@sundb 2025-04-01
@minchopaskal should we consider using SIMD to optimize these new commands?