@minchopaskal 2025-04-01
> @minchopaskal should we consider using SIMD to optimize these new commands?

@sundb thank you for the suggestion. Actually, I though about that too but it seems the code is written in such way so that the compiler specifically can generate SIMD code. I checked the disassembly of arm64 and x86 which both generate simd instructions. The new commands have implementations for the optimized path.
Maybe if performance become an issue we could look into hand-written SIMD, what do you think?