@sundb 2025-04-01
@minchopaskal In https://github.com/redis/redis/issues/13558, we found that even in the case of support SIMD, the compiler will not generate the corresponding code, maybe we can try to take a command to verify it.